Refined structure and refined logic for work with non-double values

+++ SortingAlgorithms-Rework/SortingAlgorithms/Models/Tuples/SortablePair.cs
@@ -7,9 +7,9 @@ namespace SortingAlgorithms.Models.Tuples
 {
     public struct SortablePair<T> where T : IConvertible, IComparable
     {
-        public double subv;
+        public decimal subv;
         public T value;
-        public SortablePair(double subv, T value)
+        public SortablePair(decimal subv, T value)
         {
             this.subv = subv;
             this.value = value;
